requirements.txt
aniso8601==4.0.1
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycparser==2.19
PyNaCl==1.3.0
pytz==2018.7
six==1.11.0
SQLAlchemy==1.2.13
Werkzeug==0.14.1
